+++ src/main/java/mx/com/lestradam/algorithms/functions/fitness/FFParticleSwarmOptimization.java
@@ -27,7 +27,7 @@ public class FFParticleSwarmOptimization {
 		return objFunction.evaluate(solution);
 	}
 	
-	public double excess(long[] solution) {
+	public long excess(long[] solution) {
 		return penalty.excess(solution);
 	}
