+++ src/main/java/mx/com/lestradam/algorithms/functions/fitness/PenaltyFunction.java
@@ -3,6 +3,6 @@ package mx.com.lestradam.algorithms.functions.fitness;
 public interface PenaltyFunction {
 	
 	public double evaluate(long[] solution);
-	public double excess(long[] solution);
+	public long excess(long[] solution);
 
 }
